+++ src/workbench/mod.rs
@@ -7,6 +7,7 @@ mod agent_timeline;
 mod browser_tab;
 mod chat_markdown;
 mod create_workspace_wizard;
+mod harness_chords;
 mod harness_ui;
 mod harness_voice_pane;
 mod memory_graph;